To Part 7

+++ FirstRESTServer/Models/Person.cs
@@ -10,7 +10,7 @@ namespace FirstRESTServer.Models
         public int ID { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public double PayRate { get; set; }
+        public decimal PayRate { get; set; }
         public DateTime StartDate { get; set; }
     }
 }
